+++ src/php/wishlist.php
@@ -2,13 +2,14 @@
 
     include_once("config.php");
     $book_id = $_GET['id'];
+    $user_id = $_GET['usr_id'];
     $is_include = false;
 
-    $fetch = "SELECT * FROM wishlist";
+    $fetch = "SELECT * FROM wishlist WHERE user_id=$user_id";
     $fetch_query = mysqli_query($con, $fetch);
 
     if(mysqli_num_rows($fetch_query) == 0){
-        $append = "INSERT INTO wishlist (book_id) VALUES ($book_id)";
+        $append = "INSERT INTO wishlist (book_id, user_id) VALUES ($book_id, $user_id)";
             $append_query = mysqli_query($con, $append);
             if($append_query){
                 header("location:../template/shop.php");
@@ -20,13 +21,13 @@
             }
         }
         if($is_include){
-            $remove = "DELETE FROM wishlist WHERE book_id = $book_id";
+            $remove = "DELETE FROM wishlist WHERE book_id = $book_id and user_id = $user_id";
             $remove_query = mysqli_query($con, $remove);
             if($remove_query){
                 header("location: ../template/shop.php");
             }
         }else{
-            $append = "INSERT INTO wishlist (book_id) VALUES ($book_id)";
+            $append = "INSERT INTO wishlist (book_id, user_id) VALUES ($book_id, $user_id)";
             $append_query = mysqli_query($con, $append);
             if($append_query){
                 header("location: ../template/shop.php");
